normal distobution

+++ math/0x03-probability/normal.py
@@ -8,23 +8,21 @@ distribution for a real-valued random variable.
 """
 
 
-
 class Normal:
     """Class of Normal."""
-
 
     def __init__(self, data=None, mean=0., stddev=1.):
         """Initialize."""
         if data is None:
             if stddev <= 0:
                 raise ValueError('stddev must be a positive value')
-            self.mean = mean
-            self.stddev = stddev
+            self.mean = float(mean)
+            self.stddev = float(stddev)
         else:
             if type(data) is not list:
                 raise TypeError('data must be a list')
             elif len(data) < 2:
                 raise ValueError('data must contain multiple values')
-            self.mean = sum(data) / len(data)
-            self.stddev = (sum([(x - self.mean)**2 for x in data])
-                           / len(data)) ** 1/2
+            self.mean = float(sum(data) / len(data))
+            self.stddev = float((sum([(x - self.mean)**2 for x in data])
+                                / len(data)) ** 1/2)
